+++ src/webviews/apps/shared/components/onboarding/onboarding-types.ts
@@ -1,0 +1,12 @@
+export interface OnboardingStateTemplate {
+	[key: `${string}Checked`]: boolean;
+}
+
+export interface OnboardingItemConfiguration<T extends string> {
+	itemId: T;
+	playHref?: string;
+	infoHref?: string;
+	infoTooltip?: string;
+	title: string;
+	children?: OnboardingItemConfiguration<T>[];
+}